web-animations/animation-sequence/animation-sequence-constructor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationSequence constructor test</title>
<meta name="assert" content="Creates a new AnimationSequence object">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationsequence-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var sequence = new AnimationSequence(null);

    assert_array_equals(sequence.children, [], 'AnimationSequence should have no children');
    assert_object_equals(sequence.timing, DEFAULT_TIMING, 'AnimationSequence timing members ' +
        'should be set to default values');
}, 'Test AnimationSequence(null) constructor creates AnimationSequence instance ' +
    'without children and with default timing');

test(function() {
    var sequence = new AnimationSequence([]);

    assert_array_equals(sequence.children, [], 'AnimationSequence should have no children');
    assert_object_equals(sequence.timing, DEFAULT_TIMING, 'AnimationSequence timing members ' +
        'should be set to default values');
}, 'Test AnimationSequence constructor with emplty child list and with no second argument ' +
    'creates AnimationSequence instance without children and with default timing');

test(function() {
    var sequence = new AnimationSequence([], 5);

    assert_array_equals(sequence.children, [], 'AnimationSequence should have no children');
    var timing = Object.create(DEFAULT_TIMING);
    timing.duration = 5;
    assert_object_equals(sequence.timing, timing, 'AnimationSequence timing members ' +
        'should be set to default values, except duration');
}, 'Test AnimationSequence constructor with emplty child list and with double second ' +
    'argument creates AnimationSequence instance without children and with default timing, ' +
    'except duration');

test(function() {
    var timing = {
        delay: 1,
        endDelay: 1,
        fill: 'none',
        iterationStart: 1.1,
        iterations: 2.0,
        duration: 3,
        playbackRate: -2,
        playbackDirection: 'reverse',
        direction: 'reverse',
        easing: 'ease'};
    var node = new AnimationSequence([], timing);

    assert_array_equals(node.children, [], 'AnimationSequence should have no children');
    assert_object_equals(node.timing, timing, 'AnimationSequence timing should be equal ' +
        'to the value passed to the constructor');
}, 'Test AnimationSequence constructor with emplty child list and with AnimationTimingInput ' +
    'second argument creates AnimationSequence instance without children and with given timing');

test(function() {
    var node = newAnimation(createDiv(this));
    var children = [node];
    var sequence = new AnimationSequence(children);

    assert_array_equals(sequence.children, children,
        'AnimationSequence should have a single child');
}, 'Test AnimationSequence constructor creates AnimationSequence instance with a single child');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationSequence([]);
    var node3 = newAnimation(createDiv(this));
    var children = [node1, node2, node3];
    var sequence = new AnimationSequence(children);

    assert_array_equals(sequence.children, children, 'AnimationSequence.children should ' +
        'correspond the first constructor argument');
}, 'Test AnimationSequence constructor creates AnimationSequence instance with children');

test(function() {
    var node = newAnimation(createDiv(this));
    var sequence = new AnimationSequence([node, node]);

    assert_array_equals(sequence.children, [node], 'AnimationSequence constructor should ' +
        'remove node from its parent group before inserting to child list');
}, 'Test that AnimationSequence constructor removes Animation node from its parent ' +
    'before inserting to child list even this is the same parent');

test(function() {
    var node = new AnimationGroup([]);
    var sequence = new AnimationSequence([node, node]);

    assert_array_equals(sequence.children, [node], 'AnimationSequence constructor should ' +
        'remove node from its parent group before inserting to child list');
}, 'Test that AnimationSequence constructor removes AnimationGroup node from its parent ' +
    'before inserting to child list even this is the same parent');

test(function() {
    var node = new AnimationSequence([]);
    var sequence = new AnimationSequence([node, node]);

    assert_array_equals(sequence.children, [node], 'AnimationSequence constructor should ' +
        'remove node from its parent group before inserting to child list');
}, 'Test AnimationSequence constructor removes AnimationSequence node from its parent ' +
    'before inserting to child list even this is the same parent');

test(function() {
    var node = newAnimation(createDiv(this));
    var group = new AnimationGroup([node]);
    var sequence = new AnimationSequence([node]);

    assert_array_equals(group.children, [], 'AnimationSequence constructor should ' +
        'remove node from its parent group before inserting to child list');
    assert_array_equals(sequence.children, [node], 'AnimationSequence constructor should ' +
        'insert node to the child list after removing it from its previous group');
}, 'Test AnimationSequence constructor removes Animation node from its parent ' +
    'before inserting to child list');

test(function() {
    var node = new AnimationGroup([]);
    var group = new AnimationGroup([node]);
    var sequence = new AnimationSequence([node]);

    assert_array_equals(group.children, [], 'AnimationSequence constructor should ' +
        'remove node from its parent group before inserting to child list');
    assert_array_equals(sequence.children, [node], 'AnimationSequence constructor should ' +
        'insert node to the child list after removing it from its previous group');
}, 'Test AnimationSequence constructor removes AnimationGroup node from its parent ' +
    'before inserting to child list');

test(function() {
    var node = new AnimationSequence([]);
    var group = new AnimationGroup([node]);
    var sequence = new AnimationSequence([node]);

    assert_array_equals(group.children, [], 'AnimationSequence constructor should ' +
        'remove node from its parent group before inserting to child list');
    assert_array_equals(sequence.children, [node], 'AnimationSequence constructor should ' +
        'insert node to the child list after removing it from its previous group');
}, 'Test AnimationSequence constructor removes AnimationSequence node from its parent ' +
    'before inserting to child list');
</script>
</body>
